Fix broken Padauk build
The [8f71a8c] change to Fea.pm broke Padauk font build. In the wscript, the `--markattach` parameter to `make_fea` already had the 'c' prefixed on the classname, which was then being changed to 'cc'.  So now we will assume the make_fea parameter already uses the author's desired class naming convention.